Fix #150
When calculating the height for PrefixListLayout, we were only adding up the heights of each item. When the items is empty or small, the reported height in the final layout would be too small, causing composables under it to draw over the list layout and prefixes.

The fix is to also calculate the total height of prefixes and compare with the items height. The bigger of them should be reported as the height of the list.