Enforce minimum image size for remote images
This change ensures that remote images rendered using the Coil library have a minimum size of 16.dp.

If a scaled image is smaller than this minimum, it will be displayed at the minimum size.
Additionally, if the image's dimensions after scaling are smaller than the minimum, `ContentScale.FillWidth` will be used to prevent excessive stretching or shrinking.